Missing lock in tx.go
There is a missing lock in tx.go , after the transaction is done via the RPC , the wallet is not locked hence its kept open to be used without password. If the RPC is exposed an attacker can make tx request without password as the wallet is unlocked. It is recommended to put defer lock , after the tx is executed.